Refactor/api auth 패키지 리팩토링 (#256)
1. 클래스를 조금 더 쉽게 찾기 위해 역할에 알맞게 패키지 이름 선언
