+++ xuni-api/src/main/java/com/xuni/api/auth/annotation/AuthenticatedMember.java
@@ -1,4 +1,4 @@
-package com.xuni.api.auth.presentation;
+package com.xuni.api.auth.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,6 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Admin {
+@Target(ElementType.PARAMETER)
+public @interface AuthenticatedMember {
+
 }
