+++ xuni-api/src/main/java/com/xuni/api/auth/annotation/OptionalAuthentication.java
@@ -1,4 +1,4 @@
-package com.xuni.api.auth.presentation;
+package com.xuni.api.auth.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
