+++ xuni-api/src/main/java/com/xuni/api/auth/application/jwt/JwtTokenManager.java
@@ -1,4 +1,4 @@
-package com.xuni.api.auth.support;
+package com.xuni.api.auth.application.jwt;
 
 import com.xuni.api.auth.application.MemberDetails;
 import com.xuni.api.auth.application.SimpleMemberDetails;
