+++ xuni-api/src/main/java/com/xuni/api/group/presentation/GroupCreateController.java
@@ -2,7 +2,7 @@ package com.xuni.api.group.presentation;
 
 import com.xuni.api.group.application.GroupCreateService;
 import com.xuni.api.auth.application.MemberDetails;
-import com.xuni.api.auth.presentation.AuthenticatedMember;
+import com.xuni.api.auth.annotation.AuthenticatedMember;
 import com.xuni.core.common.http.SimpleResponse;
 import com.xuni.api.group.dto.request.GroupCreateForm;
 import lombok.RequiredArgsConstructor;
