+++ xuni-api/src/main/java/com/xuni/api/review/presentation/ReviewController.java
@@ -1,7 +1,7 @@
 package com.xuni.api.review.presentation;
 
 import com.xuni.api.auth.application.MemberDetails;
-import com.xuni.api.auth.presentation.AuthenticatedMember;
+import com.xuni.api.auth.annotation.AuthenticatedMember;
 import com.xuni.core.common.http.DataResponse;
 import com.xuni.core.common.http.SimpleResponse;
 import com.xuni.api.review.application.ReviewService;
