+++ xuni-api/src/test/java/com/xuni/api/auth/application/jwt/JwtTokenProviderTest.java
@@ -1,7 +1,8 @@
-package com.xuni.api.auth.support;
+package com.xuni.api.auth.application.jwt;
 
 import com.xuni.api.auth.application.MemberDetails;
 import com.xuni.api.auth.application.SimpleMemberDetails;
+import com.xuni.api.auth.application.jwt.JwtTokenProvider;
 import com.xuni.api.support.ServiceOnlyTest;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
