+++ xuni-api/src/test/java/com/xuni/api/auth/presentation/ApiGatewayControllerTest.java
@@ -6,8 +6,8 @@ import com.xuni.api.auth.application.SimpleMemberDetails;
 import com.xuni.api.support.ControllerCommon;
 import com.xuni.api.support.JwtTestConfiguration;
 import com.xuni.core.auth.domain.Authority;
-import com.xuni.api.auth.support.JwtTokenManager;
-import com.xuni.api.auth.support.JwtTokenProvider;
+import com.xuni.api.auth.application.jwt.JwtTokenManager;
+import com.xuni.api.auth.application.jwt.JwtTokenProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
